% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphafunc.R
\name{alphafunc}
\alias{alphafunc}
\title{This gives estimation of alpha}
\usage{
alphafunc(data, country)
}
\arguments{
\item{data}{data of A,R,D}

\item{country}{a list include betas estimation based on data, infect,
populationdynamic, travelin_compartments, travelout_compartments}
}
\description{
This gives estimation of alpha
}
