% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datas_3travel-data.R
\docType{data}
\name{datas_3travel}
\alias{datas_3travel}
\title{datas_3travel is a list including 19 different simulated data sets of 3 countries
 with traffic data as in travelout_3dat.
 The six states of countries 1,2,3 are corresponding to columns 1-6,7-12,13-18, respectively.}
\format{
A list
}
\usage{
datas_3travel
}
\description{
datas_3travel is a list including 19 different simulated data sets of 3 countries
 with traffic data as in travelout_3dat.
 The six states of countries 1,2,3 are corresponding to columns 1-6,7-12,13-18, respectively.
}
\examples{
datas_3travel[[1]]
}
\keyword{datasets}
