% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{descriptive}
\alias{descriptive}
\title{This function gives descriptive statistic of a numeric vector.}
\usage{
descriptive(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
This function gives descriptive statistic of a numeric vector.
}
\examples{
x = 1:9
descriptive(x)
\dontrun{A tibble: 1 x 9
len  mean    sd   min   max   med firstpt thirdpt   IQR
<int> <dbl> <dbl> <int> <int> <int>   <dbl>   <dbl> <dbl>
 9     5  2.74     1     9     5       3       7     4
 }
}
