% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministic_postquarantine_separate.R
\name{deterministic_postquarantine_separate}
\alias{deterministic_postquarantine_separate}
\title{This function gives deterministic compartments of the evolving process during
 the quarantine period. Outputs include travelers status done
 quarantine and new active confirmed each day during the quarantine period.}
\usage{
deterministic_postquarantine_separate(theta, inp)
}
\arguments{
\item{theta}{parameter}

\item{inp}{is a list include durationquarantine : number of days and ini: initial compartments of arrival}
}
\value{
The average status of arrivals compartments after complete quarantine
}
\description{
This function gives deterministic compartments of the evolving process during
 the quarantine period. Outputs include travelers status done
 quarantine and new active confirmed each day during the quarantine period.
}
\examples{
\dontrun{## Initial Condition
S1 = 900
I1 = 100
A1 = 0
x1 = c(S1,I1,A1,0,0,0)#
days= 14
inp = list(durationquarantine = days, ini = x1)
theta0 = c(0,0,1/14,3/100,1,.5)
deterministic_postquarantine_separate(theta0,inp)}
}
