% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detfunc_1country.R
\name{detfunc_1country}
\alias{detfunc_1country}
\title{This gives deterministic realization for a given parameter and initial condition}
\usage{
detfunc_1country(theta, inp)
}
\arguments{
\item{theta}{parameter}

\item{inp}{is a list include duration : number of days and ini: initial compartments of the country}
}
\value{
The average realization of the country during the period
}
\description{
This gives deterministic realization for a given parameter and initial condition
}
