% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detfunc_postquarantine.R
\name{detfunc_postquarantine}
\alias{detfunc_postquarantine}
\title{This function gives deterministic compartments after completed quarantine for a given parameter and an initial compartments}
\usage{
detfunc_postquarantine(theta, inp)
}
\arguments{
\item{theta}{parameter}

\item{inp}{is a list include durationquarantine : number of days and ini: initial compartments of arrival}
}
\value{
The average status of arrivals compartments after complete quarantine
}
\description{
This function gives deterministic compartments after completed quarantine for a given parameter and an initial compartments
}
\examples{
\dontrun{## Initial Condition
S1 = 900
I1 = 100
A1 = 0
x1 = c(S1,I1,A1,0,0,0)#
days= 14
inp = list(durationquarantine = days, ini = x1)
theta0 = c(0,0,1/14,3/100,1,.5)
detfunc_postquarantine(theta0,inp)}
}
