% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detfunc_quarantine.R
\name{detfunc_quarantine}
\alias{detfunc_quarantine}
\title{This function gives deterministic compartments at the last day quarantine for a given parameter and an initial condition}
\usage{
detfunc_quarantine(theta, inp)
}
\arguments{
\item{theta}{parameter}

\item{inp}{is a list include durationquarantine : number of days and ini: initial compartments of the country}
}
\value{
The average realization of the country during the period
}
\description{
This function gives deterministic compartments at the last day quarantine for a given parameter and an initial condition
}
\examples{
\dontrun{## Initial Condition
S1 = 900
I1 = 100
A1 = 0
x1 = c(S1,I1,A1,0,0,0)#
days= 14
inp = list(durationquarantine = days, ini = x1)
theta0 = c(0,0,1/14,3/100,1,.5)
detfunc_quarantine(theta0,inp)}
}
