% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detfunc_travel3.R
\name{detfunc_travel3}
\alias{detfunc_travel3}
\title{This gives deterministic realization for 3 countries}
\usage{
detfunc_travel3(ini, thetas, travel, d)
}
\arguments{
\item{ini}{initial conditions}

\item{thetas}{parameter}

\item{travel}{travel data}

\item{d}{duration travel}
}
\description{
This gives deterministic realization for 3 countries
}
