% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisboxfunc.R
\name{hisboxfunc}
\alias{hisboxfunc}
\title{This function plots our data in both histogram and boxplot.}
\usage{
hisboxfunc(values)
}
\arguments{
\item{values}{numeric vector}
}
\description{
This function plots our data in both histogram and boxplot.
}
\examples{
\dontrun{#' x = 1:9
hisboxfunc(x) }
}
