% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infectfunc.R
\name{infectfunc}
\alias{infectfunc}
\title{This function estimate number infected in each country for a given data set
 under our model assumption.}
\usage{
infectfunc(data, x)
}
\arguments{
\item{data}{A (active confirmed), R(Recovered confirmed), D(Confirmed Deceased)}

\item{x}{six initial states}
}
\value{
a sequence of estimating values of delta
}
\description{
This function estimate number infected in each country for a given data set
 under our model assumption.
}
\examples{
\dontrun{
library(CONETTravel)
data = datas_3travel[[1]][,3:5]
x = c(9999620, 250,130,0,0,0)
infectfunc(data, x)
}
}
