% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popfunc.R
\name{popfunc}
\alias{popfunc}
\title{This gives estimation for the dynamic population}
\usage{
popfunc(country)
}
\arguments{
\item{country}{list include total travel in and out each day total_in,
total_out, duration nrep, population P}
}
\description{
This gives estimation for the dynamic population
}
