% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoverfunc.R
\name{recoverfunc}
\alias{recoverfunc}
\title{This function gives the average realization of 1 given country for a given
A (active confirmed), R(Recovered confirmed), D(Confirmed Deceased) sequences.}
\usage{
recoverfunc(country, theta)
}
\arguments{
\item{country}{is a list include a data matrix with A (active confirmed), R(Recovered confirmed),
D(Confirmed Deceased), populationdynamic,and infect}

\item{theta}{parameter}
}
\description{
This function gives the average realization of 1 given country for a given
A (active confirmed), R(Recovered confirmed), D(Confirmed Deceased) sequences.
}
\examples{
\dontrun{
library(CONETTravel)
library(rlist)
data = datas_3travel[[1]][,3:5]
betas = betafunc(data)
x = c(9999620, 250,130,0,0,0)
P= sum(x)
durationtravel = 84
w21= travelout_3dat[1,1]/(travelout_3dat[1,1] + travelout_3dat[1,3])
w31= travelout_3dat[1,1]/(travelout_3dat[1,1] + travelout_3dat[1,2])
out21 = travelout_3dat[,2]
out31 = travelout_3dat[,3]
total_in = out21*w21 + out31*w31
total_out = travelout_3dat[,1]
country = list(P =P, total_in = total_in, total_out= total_out,durationtravel=durationtravel)
populationdynamic = populationdynamicfunc(country)
infect = c(infectfunc(data, x),0)
country =list.append(country, data=data, populationdynamic=populationdynamic,
                    infect= infect)
                    theta = thetas_3travel[[1]][1:6]
                    recoverfunc(country,theta)
}
}
