% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_marginalestimate.R
\name{stochastic_marginalestimate}
\alias{stochastic_marginalestimate}
\title{This function helps to perform the marginal estimate for one country under the global model.}
\usage{
stochastic_marginalestimate(theta, country)
}
\arguments{
\item{theta}{parameter}

\item{country}{is a list include durationtravel : number of days, travel in compartments each day,
total_out each day, and x_ini: initial compartments of the country}
}
\value{
The realization of the country during the period take into account of travel
}
\description{
This function helps to perform the marginal estimate for one country under the global model.
}
