% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_postquarantine.R
\name{stochastic_postquarantine}
\alias{stochastic_postquarantine}
\title{This function gives stochastic compartments after completed quarantine.}
\usage{
stochastic_postquarantine(theta, inp)
}
\arguments{
\item{theta}{parameter}

\item{inp}{is a list include durationquarantine : number of days and ini: initial compartments of arrivals}
}
\value{
The stochastic status of arrivals compartments after complete quarantine
}
\description{
This function gives stochastic compartments after completed quarantine.
}
\examples{
\dontrun{## Initial Condition
S1 = 900
I1 = 100
A1 = 0
x1 = c(S1,I1,A1,0,0,0)#
days= 14
inp = list(durationquarantine = days, ini = x1)
theta0 = c(0,0,1/14,3/100,1,.5)
stochastic_postquarantine(theta0,inp)}
}
