% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_postquarantine_separate.R
\name{stochastic_postquarantine_separate}
\alias{stochastic_postquarantine_separate}
\title{This function gives stochastic compartments evolved process during the
quarantine period.
Outputs include travelers status done
 quarantine and new active confirmed during the quarantine period.}
\usage{
stochastic_postquarantine_separate(theta, inp)
}
\arguments{
\item{theta}{parameter}

\item{inp}{is a list include durationquarantine : number of days and ini: initial compartments of arrivals}
}
\value{
The stochastic status of arrivals compartments after complete quarantine
}
\description{
This function gives stochastic compartments evolved process during the
quarantine period.
Outputs include travelers status done
 quarantine and new active confirmed during the quarantine period.
}
\examples{
\dontrun{## Initial Condition
S1 = 900
I1 = 100
A1 = 0
x1 = c(S1,I1,A1,0,0,0)#
days= 14
inp = list(durationquarantine = days, ini = x1)
theta0 = c(0,0,1/14,3/100,1,.5)
stochastic_postquarantine_separate(theta0,inp)}
}
