% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_postquarantine_separate.R
\name{stochastic_postquarantine_wholeprocess}
\alias{stochastic_postquarantine_wholeprocess}
\title{This function gives stochastic compartments of the evolve process during the quarantine time
for a given parameter and an initial compartments, where R, D are the number recover
 and death each day, not a cumulative as used to be}
\usage{
stochastic_postquarantine_wholeprocess(theta, inp)
}
\arguments{
\item{theta}{parameter}

\item{inp}{is a list include durationquarantine : number of days and ini: initial compartments of arrivals}
}
\value{
The stochastic status of arrivals compartments after complete quarantine
}
\description{
This function gives stochastic compartments of the evolve process during the quarantine time
for a given parameter and an initial compartments, where R, D are the number recover
 and death each day, not a cumulative as used to be
}
\examples{
\dontrun{## Initial Condition
S1 = 900
I1 = 100
A1 = 0
x1 = c(S1,I1,A1,0,0,0)#
days= 14
inp = list(durationquarantine = days, ini = x1)
theta0 = c(0,0,1/14,3/100,1,.5)
stochastic_postquarantine_wholeprocess(theta0,inp)}
}
