% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasticmodel_1country.R
\name{stochasticmodel_1country}
\alias{stochasticmodel_1country}
\title{This function gives a stochastic realization for a given parameter and
initial condition.}
\usage{
stochasticmodel_1country(theta, inp)
}
\arguments{
\item{theta}{parameter}

\item{inp}{is a list include duration : number of days and ini: initial compartments of the country}
}
\value{
The average realization of the country during the period
}
\description{
This function gives a stochastic realization for a given parameter and
initial condition.
}
\examples{
\dontrun{## Initial Condition
P1 = 10^7
I1 = 250
A1 = 130
S1 = P1 - I1 - A1
x1 = c(S1,I1,A1,0,0,0)#
days= 84
inp = list(duration =days, ini = x1)
k= 1
theta0 = as.numeric(thetas_3travel[[k]][1:6])
stochasticmodel_1country(theta0,inp)}
}
