% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stofunc_marginalestimate.R
\name{stofunc_marginalestimate}
\alias{stofunc_marginalestimate}
\title{This gives is an marginal estimate of one country under travel}
\usage{
stofunc_marginalestimate(theta, country)
}
\arguments{
\item{theta}{parameter}

\item{country}{is a list include durationtravel : number of days, travel in compartments each day,
total_out each day, and x_ini: initial compartments of the country}
}
\value{
The realization of the country during the period take into account of travel
}
\description{
This gives is an marginal estimate of one country under travel
}
