% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stofunc_travel.R
\name{stofunc_travel}
\alias{stofunc_travel}
\title{This gives stochastic realization for travel by marginal argument}
\usage{
stofunc_travel(adjtheta, inp)
}
\arguments{
\item{adjtheta}{the given parameter}

\item{inp}{a list of x_ini duration nrep, total_out, travelin}
}
\description{
This gives stochastic realization for travel by marginal argument
}
