% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stofunc_travel3.R
\name{stofunc_travel3}
\alias{stofunc_travel3}
\title{This gives stochastic realization for 3 countries}
\usage{
stofunc_travel3(ini, thetas, travel, d)
}
\arguments{
\item{ini}{initial conditions}

\item{thetas}{parameter}

\item{travel}{travel data}

\item{d}{duration travel}
}
\description{
This gives stochastic realization for 3 countries
}
