% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stofunc_travel3_rateshutdown.R
\name{stofunc_travel3_rateshutdown}
\alias{stofunc_travel3_rateshutdown}
\title{This gives stochastic realization for 3 countries with shutdown percent strategy}
\usage{
stofunc_travel3_rateshutdown(ini, thetas, travel, d, ratein)
}
\arguments{
\item{ini}{initial conditions}

\item{thetas}{parameter}

\item{travel}{travel data in total}

\item{d}{duration travel}

\item{ratein}{rate people can get in the country}
}
\description{
This gives stochastic realization for 3 countries with shutdown percent strategy
}
