% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stofunc_travel3_trafficregulated.R
\name{stofunc_travel3_trafficregulated}
\alias{stofunc_travel3_trafficregulated}
\title{This function gives stochastic realization for 3 countries with shutdown percent strategy}
\usage{
stofunc_travel3_trafficregulated(combinetheta, inp)
}
\arguments{
\item{combinetheta}{parameter}

\item{inp}{is a list include duration : durationtravel (number of days),
travelregulated: a list of matrix travel allowed from 1 country to another during the duration,
ini: initial compartments of countries)}
}
\value{
The average realization of 3 countries with travel data regulated
}
\description{
This function gives stochastic realization for 3 countries with shutdown percent strategy
}
