% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stofunc_travel_commonshutdown.R
\name{stofunc_travel_commonshutdown}
\alias{stofunc_travel_commonshutdown}
\title{This gives stochastic realization of one country for travelin not allow at common rate}
\usage{
stofunc_travel_commonshutdown(adjtheta, inp)
}
\arguments{
\item{adjtheta}{the given parameter}

\item{inp}{a list of x_ini duration nrep, total_out, travelin, shutdowninrate, shutdownoutrate}
}
\description{
This gives stochastic realization of one country for travelin not allow at common rate
}
