% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{summaryConet}
\alias{summaryConet}
\title{This gives descriptive stat}
\usage{
summaryConet(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
This gives descriptive stat
}
\examples{
x = 1:9
summaryConet(x)
\dontrun{A tibble: 1 x 9
len  mean    sd   min   max   med firstpt thirdpt   IQR
<int> <dbl> <dbl> <int> <int> <int>   <dbl>   <dbl> <dbl>
 9     5  2.74     1     9     5       3       7     4
 }
summaryConet(x)
}
