% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetas_3travel-data.R
\docType{data}
\name{thetas_3travel}
\alias{thetas_3travel}
\title{thetas_3travel is the corresponding list of underlying true thetas which generated the simulated
data sets.}
\format{
An object of class blablabla
}
\usage{
thetas_3travel
}
\description{
thetas_3travel is the corresponding list of underlying true thetas which generated the simulated
data sets.
}
\examples{
thetas_3travel
}
\keyword{datasets}
