% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totaltravelout_samerate_regulated.R
\name{totaltravelout_samerate_regulated}
\alias{totaltravelout_samerate_regulated}
\title{This function gives a list of total travel data allowed from 1 country to
 another for a given in rate.}
\usage{
totaltravelout_samerate_regulated(traveloutdat, ratein, P)
}
\arguments{
\item{traveloutdat}{is the data of travel out from each country without pandemic}

\item{ratein}{numeric values of rate let people in}

\item{P}{vector population of countries in order as travel data}
}
\value{
A list of number travelers can go from 1 country to another each day during the duration
}
\description{
This function gives a list of total travel data allowed from 1 country to
 another for a given in rate.
}
\examples{
\dontrun{
For 3 countries
P1 = 10^7
P2 = 3*10^6
P3 = 2*10^6
P = c(P1, P2, P3) #population of 3 countries
traveloutdat = travelout_3dat
ratein = 1 # policy that allows full rate of travel in
totaltravelout_samerate_regulated(traveloutdat, ratein, P)
##########################
For 2 countries
P1 = 10^7
P2 = 3*10^6
P = c(P1, P2) #population of 2 countries
traveloutdat = travelout_3dat[,1:2] #travel out of 2 countries
ratein = .5 # policy that allows full rate of travel in
totaltravelout_samerate_regulated(traveloutdat, ratein, P)
}
}
