% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travelout_3dat-data.R
\docType{data}
\name{travelout_3dat}
\alias{travelout_3dat}
\title{travelout_3dat is simulated travel data. The matrix has 84 rows and 3 columns,
 where each row is the total number daily travel out people from countries 1,2,and 3
 each column corresponding to 1 country
 (country 1- column1, country 2 -column 2, and country 3 - column 3).}
\format{
Matrix
}
\usage{
travelout_3dat
}
\description{
travelout_3dat is simulated travel data. The matrix has 84 rows and 3 columns,
 where each row is the total number daily travel out people from countries 1,2,and 3
 each column corresponding to 1 country
 (country 1- column1, country 2 -column 2, and country 3 - column 3).
}
\examples{
travelout_3dat
}
\keyword{datasets}
