% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphafunc.R
\name{alphafunc}
\alias{alphafunc}
\title{This gives estimation of alpha}
\usage{
alphafunc(data, inp)
}
\arguments{
\item{data}{data of A,R,D}

\item{inp}{a list include beta estimation based on data, infect, pop, travelin, travelout}
}
\description{
This gives estimation of alpha
}
