% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datas_3travel-data.R
\docType{data}
\name{datas_3travel}
\alias{datas_3travel}
\title{datas_3travel here
Description of your data}
\format{
An object of class blablabla
}
\usage{
data(datas_3travel)
}
\description{
datas_3travel here
Description of your data
}
\examples{
data(datas_3travel)
}
\keyword{datasets}
