% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datas_3travel-data.R
\docType{data}
\name{datas_3travel}
\alias{datas_3travel}
\title{datas_3travel is a list including 19 different simulated data sets of 3 countries.
 The six states of country 1,2,3 corresponding to column 1-6,7-12,13-18, respectively.}
\format{
An object of class blablabla
}
\usage{
data(datas_3travel)
}
\description{
datas_3travel is a list including 19 different simulated data sets of 3 countries.
 The six states of country 1,2,3 corresponding to column 1-6,7-12,13-18, respectively.
}
\examples{
data(datas_3travel)
}
\keyword{datasets}
