% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltafunc.R
\name{deltafunc}
\alias{deltafunc}
\title{This function gives a rough estimation of of the death rate delta
 and used to improve the estimation of delta.}
\usage{
deltafunc(data)
}
\arguments{
\item{data}{data of A (active confirmed), R(Recovered confirmed), D(Confirmed Deceased)}
}
\description{
This function gives a rough estimation of of the death rate delta
 and used to improve the estimation of delta.
}
