% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detfunc_travel3_trafficregulated.R
\name{detfunc_travel3_trafficregulated}
\alias{detfunc_travel3_trafficregulated}
\title{This gives deterministic realization for 3 countries with shutdown percent strategy}
\usage{
detfunc_travel3_trafficregulated(combinetheta, inp)
}
\arguments{
\item{combinetheta}{parameter}

\item{inp}{is a list include duration : durationtravel (days),
travelregulated: a list of travel allowed from 1 country to another during the duration,
ini: initial compartments of countries)}
}
\value{
The average realization of 3 countries with travel data regulated
}
\description{
This gives deterministic realization for 3 countries with shutdown percent strategy
}
