% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisboxfunc.R
\name{hisboxfunc}
\alias{hisboxfunc}
\title{This gives descriptive stat as tibble}
\usage{
hisboxfunc(values)
}
\arguments{
\item{values}{numeric vector}
}
\description{
This gives descriptive stat as tibble
}
\examples{
\dontrun{#' x = 1:9
hisboxfunc(x) }
}
