% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infectfunc.R
\name{infectfunc}
\alias{infectfunc}
\title{This estimate number infected in each country for a given data}
\usage{
infectfunc(data, x)
}
\arguments{
\item{data}{A,R,D sequence each day}

\item{x}{initial corona}
}
\description{
This estimate number infected in each country for a given data
}
