% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infectfunc.R
\name{infectfunc}
\alias{infectfunc}
\title{This function estimate number infected in each country for a given data set
 under our model assumption}
\usage{
infectfunc(data, x)
}
\arguments{
\item{data}{A (active confirmed), R(Recovered confirmed), D(Confirmed Deceased)}

\item{x}{initial corona}
}
\description{
This function estimate number infected in each country for a given data set
 under our model assumption
}
