% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationdynamic_func.R
\name{populationdynamicfunc}
\alias{populationdynamicfunc}
\title{This function gives the estimation for the dynamic population.}
\usage{
populationdynamicfunc(country)
}
\arguments{
\item{country}{list include total travel in: total_in, total travel out each day,
total_out, durationtravel, population P}
}
\description{
This function gives the estimation for the dynamic population.
}
