% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_marginalestimate.R
\name{stochastic_marginalestimate}
\alias{stochastic_marginalestimate}
\title{This function helps to perform the marginal estimate for one country under the global model.}
\usage{
stochastic_marginalestimate(theta, country)
}
\arguments{
\item{theta}{parameter}

\item{country}{is a list include durationtravel : number of days, travel in compartments each day,
total_out each day, and x_ini: initial compartments of the country}
}
\value{
The realization of the country during the period take into account of travel
}
\description{
This function helps to perform the marginal estimate for one country under the global model.
}
\examples{
\dontrun{
library(CONETTravel)
library(rlist)
data = datas_3travel[[1]][,3:5]
x = c(9999620, 250,130,0,0,0)
P= sum(x)
durationtravel = 84
w21= travelout_3dat[1,1]/(travelout_3dat[1,1] + travelout_3dat[1,3])
w31= travelout_3dat[1,1]/(travelout_3dat[1,1] + travelout_3dat[1,2])
out21 = travelout_3dat[,2]
out31 = travelout_3dat[,3]
total_in = out21*w21 + out31*w31
total_out = travelout_3dat[,1]
country = list(P =P,  total_out= total_out,durationtravel=durationtravel)
populationdynamic = populationdynamicfunc(country)
infect = c(infectfunc(data, x),0)
travelin_compartments = matrix(0,nrow=country$durationtravel,6)
S_in = round(total_in*.97,digits=0)
I_in = round(total_in*.03,digits=0)
travelin_compartments[,1] = S_in
travelin_compartments[,2] = I_in
country =list.append(country,travelin_compartments = travelin_compartments,
populationdynamic=populationdynamic,infect= infect, x_ini = x)
theta = thetas_3travel[[1]][1:6]
stochastic_marginalestimate(theta, country)
}
}
