% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetas_3travel-data.R
\docType{data}
\name{thetas_3travel}
\alias{thetas_3travel}
\title{thetas_3travel is the corresponding list of thetas truth which generated the data}
\format{
An object of class blablabla
}
\usage{
data(thetas_3travel)
}
\description{
thetas_3travel is the corresponding list of thetas truth which generated the data
}
\examples{
data(thetas_3travel)
}
\keyword{datasets}
