% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travelout_3dat-data.R
\docType{data}
\name{travelout_3dat}
\alias{travelout_3dat}
\title{travelout_3dat is a matrix of total number out from each country during the duration travel}
\format{
An object of class blablabla
}
\usage{
data(travelout_3dat)
}
\description{
travelout_3dat is a matrix of total number out from each country during the duration travel
}
\examples{
data(travelout_3dat)
}
\keyword{datasets}
