% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveloutcompfunc.R
\name{traveloutcompfunc}
\alias{traveloutcompfunc}
\title{This function gives travel out compartment in 6 categories from 1 country to another.}
\usage{
traveloutcompfunc(total, country)
}
\arguments{
\item{total}{total travelers out}

\item{country}{a list include data of A,R,D sequence, population pop, infect}
}
\description{
This function gives travel out compartment in 6 categories from 1 country to another.
}
